Refactor external grading into separate graders
