+++ lib/assessment.js
@@ -246,72 +246,6 @@ module.exports = {
     },
 
     /**
-     * Generates an object that can be passed to assessment.processGradingResult.
-     * This function can be passed a parsed results object, or it can be passed a
-     * string or buffer to attempt to parse it and mark the grading job as failed when
-     * parsing fails.
-     *
-     * @param {Object|string|Buffer} data - The grading results
-     */
-    makeGradingResult(data) {
-        if (typeof data === 'string' || Buffer.isBuffer(data)) {
-            try {
-                data = JSON.parse(data);
-            } catch (e) {
-                return {
-                    gradingId: data.job_id,
-                    grading: {
-                        score: 0,
-                        startTime: null,
-                        endTime: null,
-                        feedback: {
-                            succeeded: false,
-                        },
-                    },
-                };
-            }
-        }
-
-        if (!data.succeeded) {
-            return {
-                gradingId: data.job_id,
-                grading: {
-                    startTime: data.start_time || null,
-                    endTime: data.end_time || null,
-                    score: 0,
-                    feedback: data
-                }
-            };
-        }
-
-        // TODO: once we have better error handling in place, account for these errors
-        /*
-        if (!data.results) {
-            return callback(new Error('results.json did not contain \'results\' object.'));
-        }
-
-        if (typeof data.results.score !== 'number' || Number.isNaN(data.results.score)) {
-            return callback(new Error('Score did not exist or is not a number!'));
-        }
-        */
-
-        let score = 0.0;
-        if (data.results && typeof data.results.score === 'number' && !Number.isNaN(data.results.score)) {
-            score = data.results.score;
-        }
-
-        return {
-            gradingId: data.job_id,
-            grading: {
-                startTime: data.start_time,
-                endTime: data.end_time,
-                score: score,
-                feedback: data
-            }
-        };
-    },
-
-    /**
      * Process the result of an external grading job.
      *
      * @param {Obect} content - The grading job data to process.
