+++ lib/config.js
@@ -25,6 +25,7 @@ config.questionDefaultsDir = 'question-servers/default-calculation';
 config.secretKey = 'THIS_IS_THE_SECRET_KEY'; // override in config.json
 config.secretSlackOpsBotEndpoint = null; // override in config.json
 config.externalGradingUseAws = false;
+config.externalGradingUseBatch = false;
 config.externalGradingJobsS3Bucket = 'prairielearn.dev.grading.jobs';
 config.externalGradingResultsS3Bucket = 'prairielearn.dev.grading.results';
 config.externalGradingArchivesS3Bucket = 'prairielearn.dev.grading.archives';
