+++ lib/externalGraderEC2.js
@@ -1,0 +1,151 @@
+var ERR = require('async-stacktrace');
+const async = require('async');
+const EventEmitter = require('events');
+const fs = require('fs-extra');
+const SHA256 = require('crypto-js/sha256');
+const targz = require('tar.gz');
+const csrf = require('./csrf');
+const _ = require('lodash');
+const AWS = require('aws-sdk');
+
+const globalConfig = require('./config');
+const externalGraderCommon = require('./externalGraderCommon');
+
+
+class Grader {
+    handleGradingRequest(grading_job, submission, variant, question, course, configOverrides) {
+        const config = _.cloneDeep(globalConfig);
+        _.assign(config, configOverrides);
+
+        const emitter = new EventEmitter();
+
+        const dir = externalGraderCommon.getJobDirectory(grading_job.id);
+
+        async.series([
+            (callback) => {
+                externalGraderCommon.buildDirectory(dir, submission, question, course, callback);
+            },
+            (callback) => {
+                // Now that we've built up our directory, let's zip it up and send
+                // it off to S3
+                let tarball = new targz({}, {
+                    fromBase: true,
+                });
+
+                let tarSrc = tarball.createReadStream(dir);
+
+                const params = {
+                    Bucket: config.externalGradingJobsS3Bucket,
+                    Key: `job_${grading_job.id}.tar.gz`,
+                };
+
+                let s3Stream = require('s3-upload-stream')(new AWS.S3());
+                let upload = s3Stream.upload(params);
+
+                upload.on('error', (err) => {
+                    ERR(err, callback);
+                });
+
+                upload.on('uploaded', () => {
+                    callback(null);
+                });
+
+                tarSrc.pipe(upload);
+            },
+            (callback) => {
+                createAndRegisterECSTaskDefinition(grading_job.id, question, config, (err) => {
+                    if (ERR(err, callback)) return;
+                    callback(null);
+                });
+            },
+            (callback) => {
+                runECSTask(grading_job.id, question, config, (err) => {
+                    if (ERR(err, callback)) return;
+                    callback(null);
+                });
+            }
+        ], (err) => {
+            fs.remove(dir);
+            if (err) {
+                emitter.emit('error', err);
+            } else {
+                emitter.emit('submit');
+            }
+        });
+
+        return emitter;
+    }
+}
+
+function getTaskDefinitionFamilyForQuestion(question) {
+    // For now, this just uses the image name as the task definition name
+    // This will create a new revision for every job using a specified
+    return SHA256(question.external_grading_image).toString();
+}
+
+function createAndRegisterECSTaskDefinition(jobId, question, config, callback) {
+    const params = {
+        containerDefinitions: [
+            {
+                name: 'grade',
+                image: question.external_grading_image,
+                memory: 512,
+                cpu: 50
+            }
+        ],
+        family: getTaskDefinitionFamilyForQuestion(question),
+        taskRoleArn: config.externalGradingJobRole,
+    };
+
+    const ecs = new AWS.ECS();
+    ecs.registerTaskDefinition(params, callback);
+}
+
+function runECSTask(jobId, question, config, callback) {
+    const params = {
+        cluster: config.externalGradingECSCluster,
+        taskDefinition: getTaskDefinitionFamilyForQuestion(question),
+        overrides: {
+            containerOverrides: [
+                {
+                    name: 'grade',
+                    environment: [
+                        {
+                            name: 'JOB_ID',
+                            value: jobId.toString(),
+                        },
+                        {
+                            name: 'ENTRYPOINT',
+                            value: question.external_grading_entrypoint,
+                        },
+                        {
+                            name: 'S3_JOBS_BUCKET',
+                            value: config.externalGradingJobsS3Bucket,
+                        },
+                        {
+                            name: 'S3_RESULTS_BUCKET',
+                            value: config.externalGradingResultsS3Bucket,
+                        },
+                        {
+                            name: 'S3_ARCHIVES_BUCKET',
+                            value: config.externalGradingArchivesS3Bucket,
+                        },
+                        {
+                            name: 'WEBHOOK_URL',
+                            value: config.externalGradingWebhookUrl,
+                        },
+                        {
+                            name: 'CSRF_TOKEN',
+                            value: csrf.generateToken({url: '/pl/webhooks/grading'}, config.secretKey),
+                        },
+                    ],
+                }
+            ]
+        }
+    };
+
+    const ecs = new AWS.ECS();
+    ecs.runTask(params, callback);
+}
+
+module.exports = Grader;
