+++ lib/question.js
@@ -5,7 +5,7 @@ var path = require('path');
 var jsonStringifySafe = require('json-stringify-safe');
 var debug = require('debug')('prairielearn:' + path.basename(__filename, '.js'));
 
-var messageQueue = require('./messageQueue');
+var externalGrader = require('./externalGrader');
 var logger = require('./logger');
 var serverJobs = require('./server-jobs');
 var sqldb = require('./sqldb');
@@ -494,7 +494,7 @@ module.exports = {
             },
             (callback) => {
                 if (!data.broken && grading_job.grading_method == 'External') {
-                    messageQueue.sendToGradingQueue(grading_job, submission, variant, question, course);
+                    externalGrader.sendToGradingQueue(grading_job, submission, variant, question, course);
                     debug('_gradeVariantWithClient()', 'sent job to grading queue');
                 }
                 callback(null);
