+++ package.json
@@ -1,63 +1,64 @@
 {
-    "name": "PrairieLearn",
-    "description": "PrairieLearn online problem-solving system",
-    "version": "2.10.0",
-    "private": true,
-    "dependencies": {
-        "adm-zip": "0.4.x",
-        "archiver": "1.3.x",
-        "async": "2.1.x",
-        "async-stacktrace": "0.0.x",
-        "aws-sdk": "2.23.x",
-        "backbone": "1.3.x",
-        "body-parser": "1.15.x",
-        "cheerio": "0.22.x",
-        "cookie-parser": "1.4.x",
-        "crypto-js": "3.1.x",
-        "csv": "1.1.x",
-        "debug": "2.6.x",
-        "diff": "3.2.x",
-        "dockerode": "2.4.x",
-        "ejs": "2.5.x",
-        "express": "4.14.x",
-        "fs-extra": "2.0.x",
-        "googleapis": "19.0.x",
-        "handlebars": "4.0.x",
-        "is-my-json-valid": "2.15.x",
-        "javascript-natural-sort": "0.7.x",
-        "json-stringify-safe": "5.0.x",
-        "jju": "1.3.x",
-        "lodash": "4.17.x",
-        "mathjax": "2.7.x",
-        "mersenne": "0.0.x",
-        "mongodb": "2.2.x",
-        "mustache": "2.3.0",
-        "numeric": "1.2.x",
-        "pg": "6.1.x",
-        "request": "2.81.x",
-        "requirejs": "2.3.x",
-        "s3-upload-stream": "1.0.x",
-        "serve-favicon": "2.3.x",
-        "socket.io": "1.7.x",
-        "socket.io-client": "1.7.x",
-        "tar-fs": "1.15.x",
-        "tar.gz": "1.0.x",
-        "winston": "2.3.x"
-    },
-    "devDependencies": {
-        "chai": "3.5.x",
-        "colors": "1.1.x",
-        "coveralls": "2.13.x",
-        "eslint": "3.19.x",
-        "jsdoc": "3.5.x",
-        "mocha": "3.2.x",
-        "nyc": "10.3.x",
-        "request": "2.79.x",
-        "yargs": "8.0.1"
-    },
-    "scripts": {
-        "test": "nyc --reporter=lcov mocha tests/index.js",
-        "lint-js": "eslint --ext js \"**/*.js\"",
-        "lint-python": "python3.6 -m flake8 ./"
-    }
+  "name": "PrairieLearn",
+  "description": "PrairieLearn online problem-solving system",
+  "version": "2.10.0",
+  "private": true,
+  "dependencies": {
+    "adm-zip": "0.4.x",
+    "archiver": "1.3.x",
+    "async": "2.1.x",
+    "async-stacktrace": "0.0.x",
+    "aws-sdk": "2.23.x",
+    "backbone": "1.3.x",
+    "body-parser": "1.15.x",
+    "cheerio": "0.22.x",
+    "cookie-parser": "1.4.x",
+    "crypto-js": "3.1.x",
+    "csv": "1.1.x",
+    "debug": "2.6.x",
+    "diff": "3.2.x",
+    "dockerode": "2.4.x",
+    "ejs": "2.5.x",
+    "express": "4.14.x",
+    "fs-extra": "2.0.x",
+    "googleapis": "19.0.x",
+    "handlebars": "4.0.x",
+    "is-my-json-valid": "2.15.x",
+    "javascript-natural-sort": "0.7.x",
+    "json-stringify-safe": "5.0.x",
+    "jju": "1.3.x",
+    "lodash": "4.17.x",
+    "mathjax": "2.7.x",
+    "mersenne": "0.0.x",
+    "mongodb": "2.2.x",
+    "mustache": "2.3.0",
+    "numeric": "1.2.x",
+    "pg": "6.1.x",
+    "request": "2.81.x",
+    "requirejs": "2.3.x",
+    "s3-upload-stream": "1.0.x",
+    "serve-favicon": "2.3.x",
+    "socket.io": "1.7.x",
+    "socket.io-client": "1.7.x",
+    "tar-fs": "1.15.x",
+    "tar.gz": "1.0.x",
+    "winston": "2.3.x"
+  },
+  "devDependencies": {
+    "chai": "3.5.x",
+    "colors": "1.1.x",
+    "coveralls": "2.13.x",
+    "eslint": "3.19.x",
+    "jsdoc": "3.5.x",
+    "mocha": "3.2.x",
+    "nyc": "10.3.x",
+    "request": "2.79.x",
+    "tmp": "0.0.33",
+    "yargs": "8.0.1"
+  },
+  "scripts": {
+    "test": "nyc --reporter=lcov mocha tests/index.js",
+    "lint-js": "eslint --ext js \"**/*.js\"",
+    "lint-python": "python3.6 -m flake8 ./"
+  }
 }
